requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.47
psycopg2-binary==2.9.6
bcrypt==4.0.1
python-dotenv==0.21.0 
openai==0.27.0  
pymysql
pytest==7.3.1